[AST] Avoid computing `getValueType` in `GenericTypeParamType`'s decl constructor
Doing so can result in a request cycle for the interface type, which
results in an ErrorType, which violates the assumption that a
`GenericTypeParamDecl` always has a `GenericTypeParamType` interface
type. We already have the code in place to defer the query to the decl,
so we don't actually need to set it at all.