[cxx-interop] Improve performance of synthesized iterator conformances
When conforming a C++ iterator type to the `UnsafeCxxInputIterator` protocol, we try to instantiate `operator==` for the given type, which is one of the requirements of the protocol.

The logic that did name lookup was unintentionally running a global lookup for `==` across all of the visible Clang modules, which became a performance issue. This happened because ClangImporter was doing a name lookup in the fake `__ObjC` module instead of the actual owning Clang module for the iterator type.

rdar://162125326 / resolves https://github.com/swiftlang/swift/issues/84733